--- a/src/nf/nf_network.f90
+++ b/src/nf/nf_network.f90
@@ -254,7 +254,6 @@ module nf_network
       integer, intent(in) :: batch_size
     end subroutine
 
-
   end interface
 
 end module nf_network