As discussed, here is a draft in which I suggst to moved the optimizer  from the network level to the layer level. 

This is just a draft with an implementation for the dense layer only.

Here are the wall clock times using my dataset (with 2 hidden dense layers):

## v0.17.0
* Forward + backward: 4.79s
* Update: 4.59s

## Current PR
* Forward + backward: 4.81s
* Update: 1.40s